+++ src/main/java/com/ldtteam/structurize/storage/ClientStructurePackLoader.java
@@ -177,7 +177,7 @@ public class ClientStructurePackLoader
      *
      * @param serverStructurePacks the server structure packs.
      */
-    public static void onServerSyncAttempt(final Map<String, Integer> serverStructurePacks)
+    public static void onServerSyncAttempt(final Map<String, Double> serverStructurePacks)
     {
         new SyncSettingsToServer().sendToServer();
 
@@ -203,7 +203,7 @@ public class ClientStructurePackLoader
         {
             if (!pack.isImmutable())
             {
-                final int version = serverStructurePacks.getOrDefault(pack.getName(), -1);
+                final double version = serverStructurePacks.getOrDefault(pack.getName(), -1.0);
                 if (version == -1)
                 {
                     if (!Structurize.getConfig().getServer().allowPlayerSchematics.get())
